requirements.txt
fastapi==0.109.1
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
google-generativeai==0.3.2
SpeechRecognition==3.10.1
gTTS==2.5.1
pydantic==2.6.1
python-jose==3.3.0
pyttsx3==2.90
pyaudio==0.2.14
pydantic-settings==2.1.0